♻️ 3MXKG6 code: decompose runner prompt hotspots
