+++ packages/agentplane/src/runner/adapters/custom.ts
@@ -19,6 +19,7 @@ import {
   runnerAdapterSuccessResult,
   type RunnerAdapter,
 } from "./shared.js";
+import { buildInvocationEventData, buildRunnerExecutionArtifacts, durationMs } from "./runtime-shared.js";
 import {
   exitCodeForSignal,
   runSupervisedProcess,
@@ -202,26 +203,6 @@ function buildCustomCommand(opts: {
   ];
 }
 
-function durationMs(startedAt: string, endedAt: string): number | undefined {
-  const started = Date.parse(startedAt);
-  const ended = Date.parse(endedAt);
-  if (Number.isNaN(started) || Number.isNaN(ended)) return undefined;
-  return Math.max(0, ended - started);
-}
-
-function buildInvocationEventData(invocation: RunnerInvocation): Record<string, unknown> {
-  return {
-    executable: invocation.argv[0] ?? null,
-    argv_count: invocation.argv.length,
-    cwd: invocation.run_dir,
-    env_keys: Object.keys(invocation.env).toSorted(),
-    trace_policy: invocation.trace_policy,
-    timeout_policy: invocation.timeout_policy,
-    has_bootstrap_path:
-      typeof invocation.bootstrap_path === "string" && invocation.bootstrap_path.trim().length > 0,
-  };
-}
-
 function normalizeCustomCommand(value: RunnerCustomConfig["command"] | undefined): string[] {
   return Array.isArray(value)
     ? value.map((entry) => entry.trim()).filter((entry) => entry.length > 0)
@@ -251,19 +232,7 @@ function buildCustomArtifacts(opts: {
   source_manifest_path?: string | null;
   invalid_manifest_path?: string | null;
 }): NonNullable<RunnerResult["artifacts"]> {
-  return (
-    runnerArtifactsFromSpecs([
-      { path: opts.invocation.bundle_path, label: "bundle" },
-      { path: opts.invocation.bootstrap_path, label: "bootstrap" },
-      { path: opts.trace_artifact_path, label: "raw-trace" },
-      { path: opts.trace_archive_path, label: "raw-trace-gzip" },
-      { path: opts.stderr_artifact_path, label: "stderr-log" },
-      { path: opts.stderr_archive_path, label: "stderr-log-gzip" },
-      { path: opts.source_manifest_path, label: "source-result-manifest" },
-      { path: opts.invalid_manifest_path, label: "invalid-result-manifest" },
-      { path: opts.invocation.result_path, label: "result-manifest" },
-    ]) ?? []
-  );
+  return buildRunnerExecutionArtifacts({ ...opts, include_output_last_message: false });
 }
 
 function mergeRunnerArtifacts(
